refactor(dsh): relocate the shared steering helpers into gates/steering.ts
The root discriminator (`isRootLikeAgent`), the `parentSession` root walk
(`rootAgentOf`) and the active-iteration compass scan (`steeringCompass`)
were defined in the goal bridge and imported by the planMode bridge. Both
bridges now import the three helpers from a neutral `gates/steering.ts`
(no barrel, no cycle); the helper bodies and their private structural
views move unchanged.

Pure relocation: plan-mode-bridge and catalog specs pass with no assertion
edits, and the moved `rootAgentOf` cycle case runs from tests/steering.spec.ts.
The goal bridge keeps a temporary consumption import for its decision-point
listener.
